+++ server/src/server/controllers/employee.ts
@@ -1,6 +1,7 @@
 import express, { Request, Response } from 'express'
 import Employee, { IEmployee } from '../models/Employee'
 
+
 class EmployeeController {
 
     app: express.Application;
@@ -11,6 +12,7 @@ class EmployeeController {
 
     // Get all employees
     public getEmployees = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         const employees = await Employee.find();
         if (!employees) {
             return res.status(404).json('No Employees found');
@@ -20,6 +22,7 @@ class EmployeeController {
 
     // New employee
     public createEmployee = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         try {
             const newEmployee: IEmployee = new Employee(req.body);
             await newEmployee.save();
@@ -31,6 +34,7 @@ class EmployeeController {
 
     // Get one
     public getEmployee = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         const employee = await Employee.findById(req.params.id);
         if (!employee) {
             return res.status(404).json('No Employee found');
@@ -40,6 +44,7 @@ class EmployeeController {
 
     // Edit employee
     public editEmployee = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         const { id } = req.params;
         const employee = req.body;
         await Employee.findByIdAndUpdate(id, { $set: employee }, { new: true });
@@ -48,6 +53,7 @@ class EmployeeController {
 
     // Delete employee
     public deleteEmployee = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         await Employee.findByIdAndRemove(req.params.id);
         res.json({ status: 'Employee Deleted' });
     };
