+++ server/src/server/controllers/resource.ts
@@ -12,6 +12,7 @@ class ResourceController {
 
     // Get all resources
     public getResources = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin' && req.userRole !== 'Manager') return res.status(403).json('Forbidden Access');
         const resources = await Resource.find();
         if (!resources) {
             return res.status(404).json('No Resources found');
@@ -21,6 +22,7 @@ class ResourceController {
 
     // New resource
     public createResource = async (req: Request, res: Response) => {
+        if (req.userRole === 'Visitor') return res.status(403).json('Forbidden Access');
         try {
             const newResource: IResource = new Resource(req.body);
             await newResource.save();
@@ -32,6 +34,7 @@ class ResourceController {
 
     // Get one
     public getResource = async (req: Request, res: Response) => {
+        if (req.userRole === 'Visitor') return res.status(403).json('Forbidden Access');
         const resource = await Resource.findById(req.params.id);
         if (!resource) {
             return res.status(404).json('No Resource found');
@@ -41,6 +44,7 @@ class ResourceController {
 
     // Edit resource
     public editResource = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         const { id } = req.params;
         const resource = req.body;
         await Resource.findByIdAndUpdate(id, { $set: resource }, { new: true });
@@ -49,6 +53,7 @@ class ResourceController {
 
     // Delete resource
     public deleteResource = async (req: Request, res: Response) => {
+        if (req.userRole !== 'Admin') return res.status(403).json('Forbidden Access');
         await Resource.findByIdAndRemove(req.params.id);
         res.json({ status: 'Resource Deleted' });
     };
